+++ app/views/home.scala.html
@@ -4,6 +4,6 @@
     <div class="jumbotron">
         <h1>Welcome Home @user.firstName @user.lastName</h1>
 
-        <a href="@routes.AuthenticationController.signOut()">Sign Out</a>
+        <a href="@routes.HomeController.logout()">Log Out</a>
     </div>
 }
