--- app/views/signin.scala.html
@@ -1,31 +1,0 @@
-@import helper._
-
-@(signInForm: Form[forms.SignInForm.Data])(implicit request: MessagesRequestHeader)
-@import request.{messages => m}
-
-@main("Sign In!") {
-
-    @request.flash.get("welcome").map { msg =>
-    <div class="col-md-6 col-md-offset-3 alert">
-        <a href="#" class="close" data-dismiss="alert">&times;</a>
-        <strong>@{msg}</strong>
-    </div>
-    }
-
-    @request.flash.get("login-error").map{ msg => 
-        <div class="col-md-6 col-md-offset-3 alert alert-danger alert-error">
-        <a href="#" class="close" data-dismiss="alert">&times;</a>
-        <strong>@{msg}</strong>
-        </div>
-    }
-    <div class="jumbotron">
-    @form(action = routes.AuthenticationController.signInSubmit()) {
-        @helper.inputText(signInForm("email"), Symbol("_label") -> null, Symbol("_showConstraints") -> false, Symbol("class") -> "form-control input-md", Symbol("placeholder") -> m("email"))
-        @helper.inputPassword(signInForm("password"), Symbol("_label") -> null, Symbol("_showConstraints") -> false, Symbol("class") -> "form-control input-md", Symbol("placeholder") -> m("password"))
-        @CSRF.formField
-
-        <button id="submit" type="submit" value="submit" class="btn btn-lg btn-primary btn-block">Sign In</button>
-        <a href="@routes.AuthenticationController.signUpForm()">Sign Up</a>
-    }
-    </div>
-}
