--- test/controllers/AuthenticationControllerSpec.scala
@@ -1,9 +1,0 @@
-package controllers
-
-import org.specs2.mock.Mockito
-import play.api.test._
-
-
-class AuthenticationControllerSpec extends PlaySpecification with Mockito {
-
-}
