+++ test/controllers/SignUpControllerSpec.scala
@@ -1,0 +1,6 @@
+package controllers
+
+import org.specs2.mock.Mockito
+import play.api.test._
+
+class SignUpControllerSpec extends PlaySpecification with Mockito {}
